# server/lib/generated/models/Data.js
'use strict';

// generated by: https://github.com/adafruit/io-swagger-templates
// using the Adafruit IO Swagger docs: https://io.adafruit.com/api/docs/api.json
const Model = require('../../model');

class Data extends Model {

  constructor(properties) {
    super(properties);
  }

  static type() {
    return 'Data';
  }

  static fields() {
    return {
      id: {
        type: 'number'
      },
      value: {
        type: 'string'
      },
      position: {
        type: 'string'
      },
      feed_id: {
        type: 'number'
      },
      group_id: {
        type: 'number'
      },
      expiration: {
        type: 'string'
      },
      lat: {
        type: 'number'
      },
      lon: {
        type: 'number'
      },
      ele: {
        type: 'number'
      },
      completed_at: {
        type: 'string'
      },
      created_at: {
        type: 'string'
      },
      updated_at: {
        type: 'string'
      },
      created_epoch: {
        type: 'number'
      }
    };
  }

}

exports = module.exports = Data;