# server/lib/generated/models/Error.js
'use strict';

// generated by: https://github.com/adafruit/io-swagger-templates
// using the Adafruit IO Swagger docs: https://io.adafruit.com/api/docs/api.json
const Model = require('../../model');

class Error extends Model {

  constructor(properties) {
    super(properties);
  }

  static type() {
    return 'Error';
  }

  static fields() {
    return {
      error: {
        type: 'string'
      }
    };
  }

}

exports = module.exports = Error;