# server/lib/models/Error.js
'use strict';

const GeneratedModel = require('../generated/models/Error.js');

class Error extends GeneratedModel {

  constructor(...args) {
    super(...args);
  }

}

exports = module.exports = Error;